requirements.txt
wheel
numpy
unidiff
pandas
seaborn
scikit-learn
scipy==1.10.1
fastparquet
matplotlib
gdown
graphviz
setuptools
networkx
gensim
transformers
#dgl==1.1.3+cu121
https://data.dgl.ai/wheels/cu121/dgl-2.0.0%2Bcu121-cp310-cp310-manylinux1_x86_64.whl
torchdata==0.7.1
torchmetrics==1.3.2
pytorch_lightning
tqdm==4.66.2
tsne-torch
node2vec
fpdf
ray
ray[tune]
#pip install "ray[tune]"
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
